fix: calculate arbitrary mask of scales
